pyproject.toml
[project]
name = "math-mentor-ai-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.4.22",
    "easyocr==1.7.0",
    "faiss-cpu==1.7.4",
    "langchain==0.1.0",
    "langchain-openai==0.0.5",
    "langgraph==0.0.20",
    "numpy==1.26.3",
    "openai==1.10.0",
    "openai-whisper==20231117",
    "pandas==2.1.4",
    "pillow==10.2.0",
    "pydub==0.25.1",
    "python-dotenv==1.0.0",
    "sentence-transformers==2.3.1",
    "streamlit==1.30.0",
    "sympy==1.12",
]
